@@ -9,19 +9,22 @@
     "lint": "prettier --check \"src/**/*.{ts,js}\""
   },
   "dependencies": {
-    "grid-engine": "^2.10.4",
-    "lodash": "^4.17.21",
-    "phaser": "^3.51.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "zustand": "^3.6.5"
+    "grid-engine": "^2.15.0",
+    "lodash": "^4.17.15",
+    "phaser": "^3.55.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "zustand": "^4.0.0-rc.1"
   },
   "devDependencies": {
-    "@snowpack/plugin-typescript": "^1.1.1",
-    "@types/canvas-confetti": "^1.0.0",
-    "@types/snowpack-env": "^2.3.2",
-    "prettier": "^2.0.5",
     "snowpack": "^2.18.4",
-    "typescript": "^4.0.0"
+    "prettier": "^2.0.5",
+    "typescript": "^4.0.0",
+    "@types/react": "^18.0.9",
+    "@types/lodash": "^4.14.182",
+    "@types/react-dom": "^18.0.3",
+    "@types/snowpack-env": "^2.3.2",
+    "@types/canvas-confetti": "^1.0.0",
+    "@snowpack/plugin-typescript": "^1.1.1"
   }
 }
