@@ -1,6 +1,7 @@
 import { PLAYER_SIZE } from "../constants/game";
 import { Audios, Maps, Sprites } from "../constants/assets";
 import { Tilesets } from "../constants/assets";
+import { useUserDataStore } from "../stores/userData";
 
 export default class BootScene extends Phaser.Scene {
   text: Phaser.GameObjects.Text;
@@ -16,7 +17,9 @@ export default class BootScene extends Phaser.Scene {
   }
 
   create(): void {
-    this.scene.start("Title");
+    const userSettings = useUserDataStore.getState().settings;
+    const startScene = userSettings.general.skipIntroScreen ? "World" : "Title";
+    this.scene.start(startScene);
 
     Object.values(Audios).forEach((audio) => {
       this.sound.add(audio);
@@ -25,7 +28,7 @@ export default class BootScene extends Phaser.Scene {
     this.sound.pauseOnBlur = false;
 
     // Mute sound by default:
-    // this.sound.mute = true;
+    this.sound.mute = !userSettings.general.enableSound;
   }
 
   loadImages(): void {
@@ -38,10 +41,16 @@ export default class BootScene extends Phaser.Scene {
     this.load.image("button1", "assets/images/ui/blue_button01.png");
     this.load.image("button2", "assets/images/ui/blue_button02.png");
 
+    // Battle
+    this.load.image("battle_background", "assets/images/ui/bb_background.png");
+    this.load.image("battle_grass", "assets/images/ui/bb_grass.png");
+
+    // Tilesets
     Object.values(Tilesets).forEach((tileset) => {
       this.load.image(tileset, `assets/images/tilesets/${tileset}.png`);
     });
 
+    // Audios
     Object.values(Audios).forEach((audio) => {
       this.load.audio(audio, `assets/audio/${audio}.ogg`);
     });
